+++ tmc3/ply.cpp
@@ -404,9 +404,9 @@ ply::read(
         return false;
       }
       auto& position = cloud[pointCounter];
-      position[0] = atof(tokens[indexX].c_str()) * positionScale;
-      position[1] = atof(tokens[indexY].c_str()) * positionScale;
-      position[2] = atof(tokens[indexZ].c_str()) * positionScale;
+      position[0] = (int)(atof(tokens[indexX].c_str()) * positionScale);
+      position[1] = (int)(atof(tokens[indexY].c_str()) * positionScale);
+      position[2] = (int)(atof(tokens[indexZ].c_str()) * positionScale);
       if (cloud.hasColors()) {
         auto& color = cloud.getColor(pointCounter);
         color[0] = atoi(tokens[indexG].c_str());
@@ -423,7 +423,7 @@ ply::read(
       }
       if (cloud.hasLaserAngles()) {
         cloud.getLaserAngle(pointCounter) =
-          std::round(atof(tokens[indexLaserAngle].c_str()));
+          (int)std::round(atof(tokens[indexLaserAngle].c_str()));
       }
       ++pointCounter;
     }
@@ -439,31 +439,31 @@ ply::read(
           if (attributeInfo.byteCount == 4) {
             float x;
             ifs.read(reinterpret_cast<char*>(&x), sizeof(float));
-            position[0] = x * positionScale;
+            position[0] =(int)(x * positionScale);
           } else {
             double x;
             ifs.read(reinterpret_cast<char*>(&x), sizeof(double));
-            position[0] = x * positionScale;
+            position[0] = (int)(x * positionScale);
           }
         } else if (a == indexY) {
           if (attributeInfo.byteCount == 4) {
             float y;
             ifs.read(reinterpret_cast<char*>(&y), sizeof(float));
-            position[1] = y * positionScale;
+            position[1] = (int)(y * positionScale);
           } else {
             double y;
             ifs.read(reinterpret_cast<char*>(&y), sizeof(double));
-            position[1] = y * positionScale;
+            position[1] = (int)(y * positionScale);
           }
         } else if (a == indexZ) {
           if (attributeInfo.byteCount == 4) {
             float z;
             ifs.read(reinterpret_cast<char*>(&z), sizeof(float));
-            position[2] = z * positionScale;
+            position[2] = (int)(z * positionScale);
           } else {
             double z;
             ifs.read(reinterpret_cast<char*>(&z), sizeof(double));
-            position[2] = z * positionScale;
+            position[2] = (int)(z * positionScale);
           }
         } else if (a == indexR && attributeInfo.byteCount == 1) {
           uint8_t val8b;
